+++ Editor/InstancePool/InstancePoolDebugger.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 using UnityEditor;
 using UnityEditor.UIElements;
@@ -19,31 +20,66 @@ namespace Vertx.Utilities.Editor
 			instancePoolDebugger.Show();
 		}
 
-		private const string noElement = "None";
-
 		private Type instancePoolType;
-		private ListView listView;
-		private readonly List<string> availableTypeNames = new List<string>
-		{
-			noElement
-		};
-		private readonly List<string> availableComponentNames = new List<string>
-		{
-			noElement
-		};
-		private readonly Dictionary<string, Component> namesToComponents = new Dictionary<string, Component>();
+		private ListView keysListView, valuesListView;
 
-		[SerializeField] private string chosenPoolType;
-		private IDictionary poolDictionary;
-		private PopupField<string> componentPopup;
-		private readonly List<Component> pooledComponents = new List<Component>();
+		[SerializeField] private Component selectedKey;
+		private readonly List<Component> componentKeys = new List<Component>();
+		private readonly List<Component> componentValues = new List<Component>();
+
+		[Flags]
+		private enum RefreshMode
+		{
+			Keys = 1,
+			Values = 1 << 1,
+			All = Keys | Values
+		}
+		
+		private void Refresh(RefreshMode valuesOnly = RefreshMode.All)
+		{
+			componentKeys.Clear();
+
+			object selectedPool = null;
+
+			// Gather all pool keys
+			foreach (IComponentPoolGroup componentPoolGroup in InstancePool.s_instancePools)
+			{
+				foreach (DictionaryEntry o in componentPoolGroup.PoolGroup)
+				{
+					var key = (Component)o.Key;
+					componentKeys.Add(key);
+
+					if (selectedKey != key) continue;
+					// value = IComponentPool<TInstanceType>
+					selectedPool = o.Value;
+				}
+			}
+
+			// Refresh key list
+			if ((valuesOnly & RefreshMode.Keys) != 0)
+			{
+				RebuildList(keysListView);
+				if (selectedKey != null)
+					keysListView.selectedIndex = componentKeys.IndexOf(selectedKey);
+			}
+
+			if (selectedPool == null)
+				ClearList(valuesListView);
+			else
+			{
+				componentValues.Clear();
+				foreach (Component o in (IEnumerable)selectedPool)
+					componentValues.Add(o);
+				RebuildList(valuesListView);
+			}
+		}
 
 		private void OnEnable()
 		{
 			EditorApplication.playModeStateChanged += ChangedPlayMode;
 
 			var root = rootVisualElement;
-			//Padding
+			// Padding
 			var padding = new VisualElement
 			{
 				style = { marginBottom = 10, marginLeft = 5, marginRight = 5, marginTop = 10 }, name = "Padding"
@@ -54,67 +90,43 @@ namespace Vertx.Utilities.Editor
 
 			root.Add(new Label("Instance Pool:") { style = { unityFontStyleAndWeight = FontStyle.Bold } });
 
-			//Type Popup
-			PopupField<string> typePopup = new PopupField<string>("Type", availableTypeNames, 0);
-			typePopup.RegisterCallback<MouseDownEvent, List<string>>((evt, typeNames) =>
+			root.Add(new Button(() => Refresh())
 			{
-				typeNames.Clear();
-				typeNames.Add(noElement);
-				IEnumerable componentPools = (IEnumerable)InstancePools.GetValue(null);
-				foreach (object componentPool in componentPools)
-				{
-					Type poolType = componentPool.GetType();
-					var genericTypeArgument = poolType.GenericTypeArguments[0];
-					typeNames.Add(genericTypeArgument.Name);
-				}
-			}, availableTypeNames);
-			typePopup.RegisterValueChangedCallback(evt =>
-			{
-				availableComponentNames.Clear();
-				availableComponentNames.Add(noElement);
-				componentPopup.SetValueWithoutNotify(noElement);
-
-				if (evt.newValue == noElement)
-				{
-					chosenPoolType = null;
-					componentPopup.SetEnabled(false);
-					return;
-				}
-
-				componentPopup.SetEnabled(true);
-				chosenPoolType = evt.newValue;
+				text = "Refresh"
 			});
-			root.Add(typePopup);
 
-			//Component Popup
-			componentPopup = new PopupField<string>("Component", availableComponentNames, 0);
-			componentPopup.RegisterCallback<MouseDownEvent, InstancePoolDebugger>((evt, debugger) => debugger.RebuildComponentDropdown(), this);
-			componentPopup.RegisterValueChangedCallback(evt =>
+			// Component Key list
+			root.Add(new Label("Key"));
+			keysListView = new ListView(componentKeys, (int)EditorGUIUtility.singleLineHeight, () =>
 			{
-				pooledComponents.Clear();
-				if (!namesToComponents.TryGetValue(evt.newValue, out var component) || poolDictionary == null)
+				VisualElement entry = CreateListEntry();
+				entry.Q<ObjectField>().pickingMode = PickingMode.Ignore;
+				return entry;
+			}, (element, i) =>
+			{
+				var objectField = element.Q<ObjectField>();
+				objectField.SetValueWithoutNotify(componentKeys[i]);
+			})
+			{
+				style =
 				{
-					componentPopup.SetValueWithoutNotify(noElement);
-					RebuildComponentDropdown();
-#if UNITY_2021_2_OR_NEWER
-					listView.Rebuild();
-#else
-					listView.Refresh();
-#endif
-					return;
+					height = 50,
+					backgroundColor = new Color(0, 0, 0, 0.15f),
+					marginTop = 5
 				}
-
-				IEnumerable set = (IEnumerable)poolDictionary[component];
-				foreach (object o in set)
-					pooledComponents.Add((Component)o);
-#if UNITY_2021_2_OR_NEWER
-				listView.Rebuild();
-#else
-				listView.Refresh();
-#endif
-			});
-			componentPopup.SetEnabled(false);
-			root.Add(componentPopup);
+			};
+			keysListView.selectedIndicesChanged += indices =>
+			{
+				int[] ind = indices.ToArray();
+				if(ind.Length > 1)
+					keysListView.selectedIndex = ind.First();
+				else
+				{
+					selectedKey = componentKeys[ind[0]];
+					Refresh(RefreshMode.Values);
+				}
+			};
+			root.Add(keysListView);
 
 #if UNITY_2020_1_OR_NEWER
 			HelpBox container = new HelpBox("Data is not refreshed in realtime.", HelpBoxMessageType.Warning);
@@ -124,19 +136,11 @@ namespace Vertx.Utilities.Editor
 			root.Add(container);
 #endif
 
-			//List View
-			listView = new ListView(pooledComponents, (int)EditorGUIUtility.singleLineHeight, () =>
-			{
-				VisualElement element = new VisualElement();
-				var objectField = new ObjectField { objectType = typeof(Component) };
-				element.Add(objectField);
-				objectField.Q(className: ObjectField.selectorUssClassName).style.display = DisplayStyle.None;
-				objectField.RegisterValueChangedCallback(evt => objectField.SetValueWithoutNotify(evt.previousValue));
-				return element;
-			}, (element, i) =>
+			// List View
+			valuesListView = new ListView(componentValues, (int)EditorGUIUtility.singleLineHeight, CreateListEntry, (element, i) =>
 			{
 				var objectField = element.Q<ObjectField>();
-				objectField.SetValueWithoutNotify(pooledComponents[i]);
+				objectField.SetValueWithoutNotify(componentValues[i]);
 			})
 			{
 				style =
@@ -146,72 +150,58 @@ namespace Vertx.Utilities.Editor
 					marginTop = 5
 				}
 			};
-			root.Add(listView);
+			root.Add(valuesListView);
 
 			rootVisualElement.SetEnabled(Application.isPlaying);
+
+			VisualElement CreateListEntry()
+			{
+				var element = new VisualElement
+				{
+					pickingMode = PickingMode.Ignore
+				};
+				var objectField = new ObjectField { objectType = typeof(Component) };
+				element.Add(objectField);
+				objectField.Q(className: ObjectField.selectorUssClassName).style.display = DisplayStyle.None;
+				objectField.RegisterValueChangedCallback(evt => objectField.SetValueWithoutNotify(evt.previousValue));
+				return element;
+			}
 		}
 
 		private void OnDisable() => EditorApplication.playModeStateChanged -= ChangedPlayMode;
 
-		private void ChangedPlayMode(PlayModeStateChange obj) => rootVisualElement.SetEnabled(obj == PlayModeStateChange.EnteredPlayMode);
-
-		private void Rebind()
+		private void ChangedPlayMode(PlayModeStateChange obj)
 		{
-			if (poolDictionary == null)
+			rootVisualElement.SetEnabled(obj == PlayModeStateChange.EnteredPlayMode);
+			switch (obj)
 			{
-				chosenPoolType = null;
-				pooledComponents.Clear();
-#if UNITY_2021_2_OR_NEWER
-				listView.Rebuild();
-#else
-				listView.Refresh();
-#endif
-				return;
-			}
-
-#if UNITY_2021_2_OR_NEWER
-			listView.Rebuild();
-#else
-			listView.Refresh();
-#endif
-		}
-
-		void RebuildComponentDropdown()
-		{
-			namesToComponents.Clear();
-			availableComponentNames.Clear();
-			availableComponentNames.Add(noElement);
-
-			IEnumerable componentPools = (IEnumerable)InstancePools.GetValue(null);
-			foreach (object componentPool in componentPools)
-			{
-				Type poolType = componentPool.GetType();
-				var genericTypeArgument = poolType.GenericTypeArguments[0];
-				if (genericTypeArgument.Name != chosenPoolType) continue;
-				FieldInfo pool = componentPool.GetType().GetField("pool", BindingFlags.Instance | BindingFlags.NonPublic);
-				poolDictionary = (IDictionary)pool.GetValue(componentPool);
-				foreach (object key in poolDictionary.Keys)
-				{
-					Component component = (Component)key;
-					string originalName = component.name;
-					string componentName = originalName;
-					int index = 1;
-					while (namesToComponents.ContainsKey(componentName))
-						componentName = $"{originalName} - {index++}";
-					availableComponentNames.Add(componentName);
-					namesToComponents.Add(componentName, component);
-				}
-
-				componentPopup.SetEnabled(true);
-				return;
+				case PlayModeStateChange.EnteredPlayMode:
+					rootVisualElement.SetEnabled(true);
+					Refresh();
+					break;
+				case PlayModeStateChange.ExitingPlayMode:
+					rootVisualElement.SetEnabled(false);
+					ClearList(valuesListView);
+					ClearList(keysListView);
+					break;
+				case PlayModeStateChange.EnteredEditMode:
+				case PlayModeStateChange.ExitingEditMode:
+				default:
+					return;
 			}
 		}
 
-		#region Reflection
+		private static void RebuildList(ListView listView)
+#if UNITY_2021_2_OR_NEWER
+			=> listView.Rebuild();
+#else
+			=> listView.Refresh();
+#endif
 
-		private static FieldInfo instancePools;
-		private static FieldInfo InstancePools => instancePools ?? (instancePools = typeof(InstancePool).GetField("instancePools", BindingFlags.NonPublic | BindingFlags.Static));
-
-		#endregion
+		private static void ClearList(ListView listView)
+		{
+			listView.itemsSource?.Clear();
+			RebuildList(listView);
+		}
 	}
 }